Extract dictionary to dedicated file
