+++ src/Fixtures/en/index.js
@@ -1,4 +1,5 @@
-import { dictionary, essentials, tactics } from './theory.en';
+import { dictionary } from './dictionary.en';
+import { essentials, tactics } from './theory.en';
 import { drills, trainings, programs } from './frisbeeDrills.en';
 
 export default {
