+++ src/Fixtures/en/theory.en.js
@@ -1,297 +1,3 @@
-export const dictionary = [
-  {
-    title: 'A',
-    data: [
-      {
-        text: 'Away',
-        translation: null,
-        definition:
-          'Side of the field opposite from where your stuff is. Often used to describe the force ("Force away")',
-      },
-    ],
-  },
-  {
-    title: 'B',
-    data: [
-      {
-        text: 'Break',
-        translation: null,
-        definition:
-          'When Forcing you are making the player holding the disc throw a certain direction. A break is when your attempt fails and they throw the direction you are guarding against. A teammate may yell “no break!” if there is an opposing player wide open on the guarded side, this means try extra hard to guard that side.',
-      },
-      {
-        text: 'Brick',
-        translation: null,
-        definition:
-          'Happens when a freshly pulled disc goes out-of-bounds. The receiving team take the disc to the brick-mark which is centrally located and 20 yards in-field from the first end zone line. A brick is signified by clapping hands once over head or by calling “Brick.”',
-      },
-    ],
-  },
-  {
-    title: 'C',
-    data: [
-      {
-        text: 'Callahan',
-        translation: null,
-        definition:
-          'It is when you catch an opponents pass in his/her end zone resulting in a point for the catchers team. This rarely ever happens, ever.',
-      },
-      {
-        text: 'Clear out',
-        translation: null,
-        definition:
-          'To get out of the area where the thrower wants to pass the disc. Absolutely necessary after making an unsuccessful cut or after throwing the pass. The importance of this is often underplayed to beginners.',
-      },
-      {
-        text: 'Cup',
-        translation: null,
-        definition: 'An extremely common and highly effective defensive zone tactic in Ultimate Frisbee.',
-      },
-      {
-        text: 'Cut',
-        translation: null,
-        definition:
-          'An attempt to get free to receive the pass. Usually starting with a body fake and/or a sudden change in direction or speed.',
-      },
-      {
-        text: 'Cutter',
-        translation: null,
-        definition:
-          'The two main offensive positions in Ultimate are the handler and the cutter. The cutters job is to cut, juke, and get open for a handler to pass to them. The main job of the cutter is to catch at all costs, and then to return it to a handler and repeat.',
-      },
-    ],
-  },
-  {
-    title: 'D',
-    data: [
-      {
-        text: 'Double Team',
-        translation: null,
-        definition:
-          'When holding the disc only one player can be within 10 feet of you, usually the player stalling you. If another player is within the 10 feet you can call “double team” and the stall count will restart. Note: opposing players are allowed to run and defend teammates within the 10 feet as long as they aren’t intentionally trying to block you. ',
-      },
-      {
-        text: 'Dump',
-        translation: null,
-        definition:
-          'Player who stands behind the thrower in order to help out (must get free for an easy pass) when the offence gets in trouble.This is a short pass, usually backwards, made when there are no good options up field to reset the stall count.',
-      },
-    ],
-  },
-  {
-    title: 'E',
-    data: [
-      {
-        text: 'Endzone',
-        translation: null,
-        definition: 'Area at the either end of the pitch within which a point is scored.',
-      },
-    ],
-  },
-  {
-    title: 'F',
-    data: [
-      {
-        text: 'Fast Stall',
-        translation: null,
-        definition:
-          'We all get antsy sometimes especially in the heat of play. Each stall count is supposed to be one-second long but in the midst of battle we may count faster then a second which can be called by the thrower, known as a stallout. If the staller doesn’t agree with the throwers call, they can Contest the call causing the stall count to move back only 2 seconds. If the play is not contested the count goes back to zero. A stall count for outdoor Ultimate is 10 seconds while an indoor game is usually 7 seconds.',
-      },
-      {
-        text: 'Flick',
-        translation: null,
-        definition:
-          'To throw the disc from the right side of the body for right handed players (or from the left for left handed players). The motion is similar in some respects to the forehand in tennis.',
-      },
-      {
-        text: 'Flow',
-        translation: null,
-        definition: 'A series of quick passes to well timed cuts – should result in an easy score.',
-      },
-      {
-        text: 'Forcing',
-        translation: null,
-        definition:
-          'To make it as difficult as possible for the thrower to throw the disc in one direction (usually one side of the field) in an attempt to make (force) him/her to make a pass to the other side. See the relevant section for how and why this is done.',
-      },
-      {
-        text: 'Forehand',
-        translation: null,
-        definition:
-          'To throw the disc from the right side of the body for right handed players (or from the left for left handed players). The motion is similar in some respects to the forehand in tennis.',
-      },
-    ],
-  },
-  {
-    title: 'H',
-    data: [
-      {
-        text: 'Handler',
-        translation: null,
-        definition:
-          'There are really only two main positions in Ultimate, The handler and the cutter. The handler is the one that handles the disc. Typically, the handler is an experienced player that has mastered at least the three main throws of Ultimate.. the hammer, backhand, and forehand. The handlers job is to stay back while the cutters try to get open. The handler will throw to the cutter to hopefully progress up the field. Since not everyone is capable of being a handler, the term can also used to describe a person’s skill level.',
-      },
-      {
-        text: 'Hammer',
-        translation: null,
-        definition:
-          ' An overhead throw with a forehand grip in which the frisbee is released at an angle so that it flattens out and flies upside down.',
-      },
-      {
-        text: 'Huck',
-        translation: null,
-        definition: 'A long throw down field. Huck may be spelled differently depending on the country you play.',
-      },
-    ],
-  },
-  {
-    title: 'I',
-    data: [
-      {
-        text: 'Iso',
-        translation: null,
-        definition:
-          'In this offensive strategy three of the cutters line up deeper than usual (this can vary from 5 yards farther downfield to at the end zone) while the remaining cutter lines up closer to the handlers. This closest cutter is known as the "feature," or "German." The idea behind this strategy is that it opens up space for the feature to cut, and at the same time it allows handlers to focus all of their attention on only one cutter. ',
-      },
-    ],
-  },
-  {
-    title: 'L',
-    data: [
-      {
-        text: 'Layout',
-        translation: null,
-        definition:
-          'When the player dives the catch or intercept the disc. Also referred to as “going ho” (from going horizontal).',
-      },
-    ],
-  },
-  {
-    title: 'M',
-    data: [
-      {
-        text: 'Man on Man D',
-        translation: null,
-        definition:
-          'The most common type of defence. Each person on defense marks an offence player and attempts to stay as close as possible with the intention of getting an interception or forcing a mistake.',
-      },
-      {
-        text: 'Mark',
-        translation: null,
-        definition:
-          'When you are told to cover a specific person they are your mark. Stay on your mark at all costs.  It is your job to not let that single person get the disc, and to escape them when you are trying to get open for the disc.',
-      },
-      {
-        text: 'Mismatch',
-        translation: null,
-        definition:
-          'When playing Man Defense, you will be assigned a mark to cover and to stay on throughout the entire point. Typically, you will be marking someone equal, or close to equal, your height and skill level. A Mismatch occurs when one of the two opposing players is significantly better or worse than the other. An example could be putting a 5′ 2″ guy on an opposing 6′ 5″ guy or putting a rookie player on a veteran player. Mismatches will always occur especially if the other teams skill levels are unknown.',
-      },
-    ],
-  },
-  {
-    title: 'P',
-    data: [
-      {
-        text: 'Pancake',
-        translation: null,
-        definition:
-          'A type of catch where a player claps his/her hands on the top and bottom of the disc. This is the most secure way to catch but severely limits one’s reach.',
-      },
-      {
-        text: 'Pick',
-        translation: null,
-        definition:
-          'When guarding a man on defense you are given the right to guard them without interference. So say your on defense and guarding an opposing player by basically following them everywhere they go.',
-      },
-      {
-        text: 'Pivot',
-        translation: null,
-        definition:
-          'When you plant your foot (left for right handers and right for left handers) and step to the side (allowing you the throw around the marker).',
-      },
-      {
-        text: 'Poach',
-        translation: null,
-        definition:
-          'A player would yell “Poach!” if he/she is WIDE open. Not just a little open but really open. 99% of the time the player yelling poach would receive the disc.',
-      },
-      {
-        text: 'Pull',
-        translation: null,
-        definition:
-          'When a defender moves away from their marker to try and make an interception on a pass to another player.',
-      },
-    ],
-  },
-  {
-    title: 'S',
-    data: [
-      {
-        text: 'Scoober',
-        translation: null,
-        definition:
-          'a fast, difficult-to-intercept throw; similar to a hammer, a scoober is held overhead, but with a backhand grip.',
-      },
-      {
-        text: 'Stack',
-        translation: null,
-        definition: 'A very common strategy used by many teams.',
-      },
-      {
-        text: 'Stall',
-        translation: null,
-        definition:
-          'The player holding the disc has just ten seconds to pass it to a team-mate – the defender marking the player with the disc counts to ten out loud, and if the disc has not been released on “ten” the defender takes possession. Forcing the thrower to make a less-than-ideal pass as the “stall count” nears ten is the idea behind most defensive strategies.',
-      },
-      {
-        text: 'Strike',
-        translation: null,
-        definition:
-          'When a player is forcing a particular way they are leaving a portion of the field completely open. It is your teams’ job to guard that open portion of the field. Sometimes an opposing player will stray away from a team member and become completely open. A team member would yell “Strike!” for you to change your force for about 3 seconds, long enough for them to get back on their man. Make sure you return to your normal force right after.',
-      },
-      {
-        text: 'Swing',
-        translation: null,
-        definition: ' A pass across the field to look for options on the other side of the pitch.',
-      },
-      {
-        text: 'Switch',
-        translation: null,
-        definition: 'This is when two defenders exchange the offensive players that they are marking.',
-      },
-    ],
-  },
-  {
-    title: 'T',
-    data: [
-      {
-        text: 'Travel',
-        translation: null,
-        definition:
-          'A violation which can be called when a receiver of a throw does not attempt to stop immediately or drags their planted foot before a throw. This means the disc must go back to where it was before the throw.',
-      },
-      {
-        text: 'Turn/Turnover',
-        translation: null,
-        definition:
-          'A team has possession of the disc until there is a turnover. It is how it sounds, you turn it over to the other team. Players usually call it a “turn” for short. A turn is a result of a dropped or defended pass, a stall-out, an un-contested foul, or an out-of-bounds.',
-      },
-    ],
-  },
-  {
-    title: 'Z',
-    data: [
-      {
-        text: 'Zone',
-        translation: null,
-        definition:
-          'Defensive strategy where defensive players mark zones or areas of the pitch trying to prevent space to pass into. Usually has a ‘wall’ or a ‘cup’ formation.',
-      },
-    ],
-  },
-];
-
 export const essentials = [
   {
     title: 'Throwing',
